added average rating per day endpoint

+++ backend/src/main/java/de/bigabig/wikihowqa/model/DatePoint.java
@@ -8,12 +8,12 @@ public class DatePoint {
 
     @JsonFormat(pattern="yyyy-MM-dd HH:mm:ss", timezone="Europe/Berlin")
     private Date t;
-    private float y;
+    private double y;
 
     public DatePoint() {
     }
 
-    public DatePoint(Date t, float y) {
+    public DatePoint(Date t, double y) {
         this.t = t;
         this.y = y;
     }
@@ -26,11 +26,11 @@ public class DatePoint {
         this.t = t;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
